ktl/workflow.py: Properties class: allow properties to be removed
We don't have any mechanism to remove already set properties. Add
support to remove any property (key) which has no value (None): when
flushing we avoid adding the property. This will be used by next change
in kteam-tools, as there are cases where sru-workflow-manager wants to
remove the kernel-stable-Promote-to-proposed-end property of a bug.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>