stable/check-for-fixes: new script to "grep" for fixes for a given commit
check-for-fixes is intended to find possible fixes for regressions or
other issues introduced by a set of commits on a given linux kernel
tree. Usually regression/stable fixes references the previous commits
by either the commit hash, subject, or both: this script looks for
hashes/subjects cited on upstream commits, and prints the resulting
output, so it can be manually checked if it's a relevant patch/fix
that could be applied to the tree.

Examples:

1. Checking if commit 06222e4 introduced problems (need fixes on top)

linux.git$ check-for-fixes --base=06222e4^..06222e4

...
Commit 06222e4 is referenced in b48c6af
Commit 06222e4 is referenced in 6a82c47

Checking commits b48c6af/6a82c47 that references 06222e4, we see that
they are fixes for problems introduced by 06222e4

2. Check possible regressions between ubuntu releases:

ubuntu-maverick.git$ git checkout master-next
ubuntu-maverick.git$ check-for-fixes --base=Ubuntu-2.6.35-28.50..Ubuntu-2.6.35-30.61

...
Commit a97590e is referenced in 9b4554b
Commit 1ddd504 is referenced in 07d0c38
Commit d199c96 is referenced in aa1b13e
...

In the listing that the script shows for this case, for example we can
see that 1ddd504, which is applied/cherry-picked between
Ubuntu-2.6.35-28.50..Ubuntu-2.6.35-30.61, has a regression, which is
fixed by 07d0c38 upstream and is not yet applied to the branch we are
currently in (master-next).

Note though that not all commits listed have regressions, we can have
"false positives" as sometimes commits can be cited in the changelog on
newer commits upstream for other reasons.

3. Check changes which are applied since previous Ubuntu release, and
are referenced on newer commits upstream (which possibly are fixes to
the changes we have applied). This is useful for example after applying
upstream stable updates, to know if we already have possible problems
which are already fixed upstream.

ubuntu-maverick.git$ git checkout master-next
ubuntu-maverick.git$ check-for-fixes
Filtering from 2083/2083 commits...

Commit 15bed0f is referenced in 3e309cd
Commit fe47ae7 is referenced in a0e3e70
Commit 0837e32 is referenced in d819437
...

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>