UBUNTU: Enable iwlwifi for compat-wireless-3.2
BugLink: http://bugs.launchpad.net/bugs/923900

Per LP:918351 we provided the v3.2 compat-wireless stack via LBM for
Oneiric. Unfortunately the iwlwifi drivers were not enabled. This is due
to the fact that in v3.2, iwlagn was renamed to iwlwifi and the
corresponding config names were also updated to reflect this re-name,
ie:

  commit 3c607d27c818cf4a5d28f2c73b18a88f8fbdfa33
  Author: Don Fry <donald.h.fry@intel.com>
  Date: Fri Sep 30 11:40:20 2011 -0700

    iwlagn: rename iwlagn module iwlwifi and alias to iwlagn.

    Rename the iwlagn module as iwlwifi in preparation for future
    changes. Add an alias to iwlagn for backward compatibility.

As a result, the v3.2 compat-wirless stack enables iwlwifi if
CONFIG_IWLWIFI. However, because CONFIG_IWLWIFI does not exist for
Oneiric, the v3.2 compat-wireless version of iwlwifi was/is never built.

Update the Makefiles to check for CONFIG_IWLAGN instead of
CONFIG_IWLWIFI.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>