UBUNTU: SAUCE: Return deny all as default in encode_share_access
This used to be catched by BUG() which is not only unnecessary.
It was changed as part of a bigger rewrite that depends on a
typedef which does not exist before 2.6.28.

CVE-2011-4324
BugLink: http://bugs.launchpad.net/bugs/917829

(very loosely backported from dc0b027dfadfcb8a5504f7d8052754bf8d501ab9)
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>